Windows, site utility and help, testFaster fix.
There were some problems installing and running on Windows because the shebang line is used by the npm installer even for Windows.  The README,md file is now rendered in HTML on the web UI, the test cache is no longer overwriting its record of the last filename, and more useful things are clickable in the web UI.  tt.js is for manual testing.